switch to CI/CD pipeline mode
