+++ wasm/packages/okapi/karma.conf.ts
@@ -13,7 +13,7 @@ export default (config: Config): void => {
     // base path that will be used to resolve all patterns (eg. files, exclude)
     basePath: "./",
 
-    singleRun: false,  // run and exit for CI pipelines, lol
+    singleRun: true,  // run and exit for CI pipelines, lol
 
     client: {
       clearContext: false, // will show the results in browser once all the testcases are loaded
@@ -40,7 +40,7 @@ export default (config: Config): void => {
       "test/**/*.spec.ts": ["webpack"],
     },
     // "Chrome", "ChromeHeadless"
-    browsers: ["Chrome"],
+    browsers: ["ChromeHeadless"],
 
     webpackMiddleware: {
       stats: "errors-only",
